llm_part/requirements.txt
dadata==24.10.0
requests==2.32.3
python-dotenv==1.1.0
beautifulsoup4==4.13.4